% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bike}
\alias{bike}
\title{The number of bikes rented..}
\format{A data frame with 731 rows and 12 variables:
\describe{
  \item{season}{...}
  \item{year}{...}
  \item{month}{...}
  \item{holiday}{...}
  \item{weekday}{...}
  \item{working_day}{...}
  \item{weather}{...}
  \item{temperature}{...}
  \item{humidity}{...}
  \item{wind_speed}{...}
  \item{bikes_rented}{...}
  \item{days_since_2011}{...}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
bike
}
\description{
A dataset containing the number of bikes rented.
}
\keyword{datasets}
