% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{calculate_mean_prediction}
\alias{calculate_mean_prediction}
\title{Mean Predict}
\usage{
calculate_mean_prediction(object, data, class)
}
\arguments{
\item{object}{An object of class `model_fit`.}

\item{data}{A rectangular data object, such as a data frame.}

\item{class}{The class probability to predict if the object has mode classification.}
}
\description{
Mean Predict
}
