% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{center_prediction}
\alias{center_prediction}
\title{Center Prediction}
\usage{
center_prediction(data)
}
\arguments{
\item{data}{A rectangular data object, such as a data frame.}
}
\description{
Center Prediction
}
