% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{dependence_data}
\alias{dependence_data}
\title{Dependence Data}
\usage{
dependence_data(object, new_data, feature_name, class)
}
\arguments{
\item{object}{object}

\item{new_data}{new_data}

\item{feature_name}{feature_name}

\item{class}{class}
}
\description{
The dependence_data function generates a data frame of predictions from new_data that
correspond to changes in the values of feature_name.
}
