% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{dependence_plot}
\alias{dependence_plot}
\title{Dependence Plot}
\usage{
dependence_plot(object, new_data, feature_name, class = 1,
  title = "Partial Dependence Plot", subtitle = NULL)
}
\arguments{
\item{object}{An object of class model_fit.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{feature_name}{A feature name.}

\item{class}{A class probability to predict.}

\item{title}{A character string for the title.}

\item{subtitle}{A character string for the subtitle displayed under the title.}
}
\description{
The dependence_plot function generates a plot of predictions from new_data that
correspond to changes in the values of feature_name.
}
