% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effect_data}
\alias{effect_data}
\title{Effect Data}
\usage{
effect_data(object, new_data, feature_name)
}
\arguments{
\item{object}{object}

\item{new_data}{new_data}

\item{feature_name}{feature_name}
}
\description{
Effect Data
}
