% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effect_plot}
\alias{effect_plot}
\title{Effect Plot}
\usage{
effect_plot(object, new_data, feature_name)
}
\arguments{
\item{object}{object}

\item{new_data}{new_data}

\item{feature_name}{feature_name}
}
\description{
Effect Plot
}
