% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{estimate_importance}
\alias{estimate_importance}
\title{Estimate Importance}
\usage{
estimate_importance(object, data, feature, target, metric, sample_size)
}
\arguments{
\item{object}{An object of class `model_fit`.}

\item{data}{A rectangular data object, such as a data frame.}

\item{feature}{A feature variable.}

\item{target}{The target variable.}

\item{metric}{A metric funtion from the `yardstick` package.}

\item{sample_size}{The sample size used to estimate mean importance.}
}
\description{
Estimate Importance
}
