% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{estimate_metric}
\alias{estimate_metric}
\title{Estimate Metric}
\usage{
estimate_metric(object, data, target, metric)
}
\arguments{
\item{object}{An object of class `model_fit`.}

\item{data}{A rectangular data object, such as a data frame.}

\item{target}{The target variable.}

\item{metric}{A metric funtion from the `yardstick` package.}
}
\description{
Estimate Metric
}
