% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{feature_replace}
\alias{feature_replace}
\title{Feature Replace}
\usage{
feature_replace(new_data, feature_name, feature_value)
}
\arguments{
\item{new_data}{A rectangular data object, such as a data frame.}

\item{feature_name}{A feature name.}

\item{feature_value}{A feature value of class numeric or factor.}
}
\description{
Feature Replace
}
