% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Importance}
\usage{
importance(object, data, target_name)
}
\arguments{
\item{object}{object}

\item{data}{data}

\item{target_name}{target_name}
}
\description{
Importance
}
