% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance_data}
\alias{importance_data}
\title{Importance Data}
\usage{
importance_data(object, new_data)
}
\arguments{
\item{object}{An object of class model_fit.}

\item{new_data}{A rectangular data object, such as a data frame.}
}
\description{
The importance_data function generates a data frame of permutation importance
values for each of the features in new_data.
}
