% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance_plot}
\alias{importance_plot}
\title{Importance Plot}
\usage{
importance_plot(object, data, target_name)
}
\arguments{
\item{object}{object}

\item{data}{data}

\item{target_name}{target_name}
}
\description{
Importance Plot
}
