% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{mean_predict}
\alias{mean_predict}
\title{Mean Predict}
\usage{
mean_predict(object, new_data, class)
}
\arguments{
\item{object}{An object of class model_fit with mode "regression" or "classification".}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{class}{A class probability to predict.}
}
\description{
Mean Predict
}
