% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{object_metric}
\alias{object_metric}
\title{Object Metric}
\usage{
object_metric(object, new_data)
}
\arguments{
\item{object}{An object of class model_fit.}

\item{new_data}{A rectangular data object, such as a data frame.}
}
\description{
Object Metric
}
