% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{permute_feature}
\alias{permute_feature}
\title{Permute Feature}
\usage{
permute_feature(data, feature)
}
\arguments{
\item{data}{A rectangular data object, such as a data frame.}

\item{feature}{A feature variable.}
}
\description{
Permute Feature
}
