% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{replace_feature}
\alias{replace_feature}
\title{Replace Replace}
\usage{
replace_feature(data, feature, feature_value)
}
\arguments{
\item{data}{A rectangular data object, such as a data frame.}

\item{feature}{A feature variable.}

\item{feature_value}{A feature variable value.}
}
\description{
Replace Replace
}
