% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{sequence_feature}
\alias{sequence_feature}
\title{Feature Sequence}
\usage{
sequence_feature(feature, len)
}
\arguments{
\item{feature}{A feature variable.}

\item{len}{The feature sequence length if the feature variable is of class `numeric`.}
}
\description{
Feature Sequence
}
