% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{dependence_plot}
\alias{dependence_plot}
\title{Dependence Plot}
\usage{
dependence_plot(object, new_data, feature_name,
  title = "Partial Dependence Plot", subtitle = NULL)
}
\arguments{
\item{object}{object}

\item{new_data}{new_data}

\item{feature_name}{feature_name}

\item{title}{title}

\item{subtitle}{subtitle}
}
\description{
The dependence_plot function generates a plot of predictions from new_data that
correspond to changes in the values of feature_name.
}
