% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{estimate_dependence}
\alias{estimate_dependence}
\title{Estimate Dependence}
\usage{
estimate_dependence(object, data, feature, len, class, center)
}
\arguments{
\item{object}{An object of class `model_fit`.}

\item{data}{A rectangular data object, such as a data frame.}

\item{feature}{A feature variable.}

\item{len}{A length of the feature sequence if the object has mode regression.}

\item{class}{The class probability to predict if the object has mode classification.}

\item{center}{Center the depedence plot on the first feature value.}
}
\description{
Estimate Dependence
}
