% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{feature_permutation}
\alias{feature_permutation}
\title{Feature Permutation}
\usage{
feature_permutation(new_data, feature_name)
}
\arguments{
\item{new_data}{A rectangular data object, such as a data frame.}

\item{feature_name}{A feature name.}
}
\description{
Feature Permutation
}
