% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{feature_permutation}
\alias{feature_permutation}
\title{Feature Permutation}
\usage{
feature_permutation(data, feature_name)
}
\arguments{
\item{data}{data}

\item{feature_name}{feature_name}
}
\description{
Feature Permutation
}
