% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{feature_replace}
\alias{feature_replace}
\title{Feature Replace}
\usage{
feature_replace(new_data, feature_name, feature_value)
}
\arguments{
\item{new_data}{new_data}

\item{feature_name}{feature_name}

\item{feature_value}{feature_value}
}
\description{
Feature Replace
}
