% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance_data}
\alias{importance_data}
\title{Importance Data}
\usage{
importance_data(object, new_data, target_name)
}
\arguments{
\item{object}{object}

\item{new_data}{new_data}

\item{target_name}{target_name}
}
\description{
Importance Data
}
