% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance_plot}
\alias{importance_plot}
\title{Importance Plot}
\usage{
importance_plot(object, new_data, title = "Permutation Importance Plot",
  subtitle = NULL)
}
\arguments{
\item{object}{An object of class model_fit.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{title}{A character string for the title.}

\item{subtitle}{A character string for the subtitle displayed under the title.}
}
\description{
The importance_plot function generates a plot of permutation importance
values for each of the features in new_data.
}
