% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance_plot}
\alias{importance_plot}
\title{Importance Plot}
\usage{
importance_plot(object, new_data, target_name,
  title = "Permutation Importance Plot", subtitle = NULL)
}
\arguments{
\item{object}{object}

\item{new_data}{new_data}

\item{target_name}{target_name}

\item{title}{title}

\item{subtitle}{subtitle}
}
\description{
The importance_plot function generates a plot of permutation importance
values for each of the features in new_data.
}
