% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{mean_predict}
\alias{mean_predict}
\title{Mean Predict}
\usage{
mean_predict(object, new_data, class)
}
\arguments{
\item{object}{object}

\item{new_data}{new_data}

\item{class}{class}
}
\description{
Mean Predict
}
