% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{object_metric}
\alias{object_metric}
\title{Object Metric}
\usage{
object_metric(object, new_data, target_name)
}
\arguments{
\item{object}{object}

\item{new_data}{new_data}

\item{target_name}{target_name}
}
\description{
Object Metric
}
