% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence.R
\name{plot_dependence}
\alias{plot_dependence}
\title{Plot Dependence}
\usage{
plot_dependence(object, data, feature, len = 40, class = 1,
  examples = FALSE, center = FALSE)
}
\arguments{
\item{object}{An object of class `model_fit`.}

\item{data}{A rectangular data object, such as a data frame.}

\item{feature}{A feature variable.}

\item{len}{The feature sequence length if the feature variable is of class `numeric`.}

\item{class}{The class probability to predict if the object has mode classification.}

\item{examples}{Display the dependence plot of all the examples.}

\item{center}{Center the depedence plot on the first feature value.}
}
\description{
Plot Dependence
}
