% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{plot_importance}
\alias{plot_importance}
\title{Plot Importance}
\usage{
plot_importance(object, data, target, metric, sample_size = 100,
  title = "Permutation Importance")
}
\arguments{
\item{object}{An object of class model_fit.}

\item{data}{A rectangular data object, such as a data frame.}

\item{target}{The target variable.}

\item{metric}{A metric funtion from the `yardstick` package.}

\item{sample_size}{The sample size used to estimate mean importance.}

\item{title}{A character string for the title.}
}
\description{
Plot Importance
}
